cambios menores
- paso código de creación de tarjetas de app.jsx a Caracteristicas.jsx
- paso estilo de app.css a index.css y borro app.css
- borro navbar.jsx (no era usado)
