--- petdocs/src/App.css
@@ -1,4 +1,0 @@
-#root {
-  max-width: 1440px;
-}
-
