+++ petdocs/src/components/Caracteristicas.jsx
@@ -1,18 +1,51 @@
 import React from "react";
 import "./css/Caracteristicas.css";
 
-const Caracteristicas = ({ cardData }) => {
+//Images
+import pathYellowCircle from "../assets/yellow_circle_.png";
+import pathOrangeCircle from "../assets/orange_circle_.png";
+import pathRedCircle from "../assets/red_circle_.png";
+import pathPaw from "../assets/paw-icon2.png";
+import pathAgenda from "../assets/icon_agenda.png";
+import pathPen from "../assets/icon_lapiz.png";
+
+let cards = [
+  {
+    id: "yellowCard",
+    title: "Añadir mascota",
+    link: "",
+    background: pathYellowCircle,
+    icon: pathPen,
+  },
+  {
+    id: "orangeCard",
+    title: "Ver mascota",
+    link: "",
+    background: pathOrangeCircle,
+    icon: pathPaw,
+  },
+  {
+    id: "redCard",
+    title: "Crear recordatorio",
+    link: "",
+    background: pathRedCircle,
+    icon: pathAgenda,
+  },
+];
+
+const Caracteristicas = () => {
   // Clases image-overlay y overlay-image no parecen hacer nada.
   return (
     <div className="cardsContainer">
-      {cardData.map((data, index) => (
+      {cards.map((data, index) => (
         <a href={data.link}>
           <div className="card" key={index}>
             <div className="image-overlay imageContainer">
               <img
                 src={data.background}
-                className={`card-img-top ${data.id === "redCard" ? 'card-back-opacity' : ''}`}
-                // className="card-img-top"
+                //Cambiar el className usado en caso que se quiera ver más claro el fondo del círculo rojo (habilitar su correspondiente estilo también).
+                // className={`card-img-top ${data.id === "redCard" ? 'card-back-opacity' : ''}`}
+                className="card-img-top"
                 alt="Background"
               />
               <div className="icon-and-text">
