--- petdocs/src/components/Navbar.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const Navbar = () => (
-  <nav>
-    <ul>
-      <li><Link to="/">Home</Link></li>
-      <li><Link to="/register">Register</Link></li>
-      <li><Link to="/login">Login</Link></li>
-    </ul>
-  </nav>
-);
-
-export default Navbar;
