+++ petdocs/src/index.css
@@ -27,4 +27,5 @@ body {
 
 #root {
   width: 100%;
+  max-width: 1440px;
 }
